backend/requirements.txt
fastapi==0.109.2
httptools==0.6.1
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
urllib3==2.2.0
uvicorn==0.27.1
websockets==12.0
